Change the name of some variables
Some variables have been renamed from $1 or $i to something more
maningful.
Some more local variables have been added to clear the global space.
The token is appended to the url in the main function now since it
didn't make lot of sense to have it in read_config().
The script has been formated with shell-format.